pyproject.toml
[project]
name = "lsnz"
version = "1.0.0"
description = "New and improved LSNZ Website"
requires-python = ">=3.12"
dependencies = [
    "flask",
    "flask-sqlalchemy",
    "flask-migrate",
    "dotenv>=0.9.9",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "email-validator>=2.2.0",
    "markdown>=3.8.2",
    "bootstrap-flask>=2.5.0",
    "flask-principal>=0.4.0",
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]
